requirements.txt
pytest
appium-python-client
allure-pytest
allure-python-commons